refactor(frontend): move WalletConnect generic constants to lib
