+++ src/frontend/src/lib/components/wallet-connect/WalletConnectDomainVerification.svelte
@@ -1,9 +1,9 @@
 <script lang="ts">
 	import type { Verify } from '@walletconnect/types';
 	import type { Web3WalletTypes } from '@walletconnect/web3wallet';
-	import { CONTEXT_VALIDATION_ISSCAM } from '$eth/constants/wallet-connect.constants';
 	import { i18n } from '$lib/stores/i18n.store';
 	import type { Option } from '$lib/types/utils';
+	import { CONTEXT_VALIDATION_ISSCAM } from '$lib/constants/wallet-connect.constants';
 
 	export let proposal: Option<Web3WalletTypes.SessionProposal>;
