+++ src/frontend/src/lib/providers/wallet-connect.providers.ts
@@ -4,9 +4,9 @@ import {
 	SESSION_REQUEST_ETH_SEND_TRANSACTION,
 	SESSION_REQUEST_ETH_SIGN,
 	SESSION_REQUEST_ETH_SIGN_V4,
-	SESSION_REQUEST_PERSONAL_SIGN,
-	WALLET_CONNECT_METADATA
+	SESSION_REQUEST_PERSONAL_SIGN
 } from '$eth/constants/wallet-connect.constants';
+import { WALLET_CONNECT_METADATA } from '$lib/constants/wallet-connect.constants';
 import type { EthAddress, OptionSolAddress } from '$lib/types/address';
 import type { WalletConnectListener } from '$lib/types/wallet-connect';
 import { Core } from '@walletconnect/core';
